add SeekBtn
